+++ LowerTriangularArrays/src/lower_triangular_array.jl
@@ -791,7 +791,7 @@ Adapt.adapt_structure(to, L::LowerTriangularArray) = Adapt.adapt(to, L.data)
 
 Architectures.architecture(L::LowerTriangularArray) = architecture(L.spectrum)
 
-function Architectures.on_architecture(arch, L::LowerTriangularArray)
+function Architectures.on_architecture(arch::AbstractArchitecture, L::LowerTriangularArray)
     adapted_data = on_architecture(arch, L.data)
     if ismatching(L.spectrum, typeof(adapted_data)) # if matching, use the same spectrum
         return LowerTriangularArray(adapted_data, on_architecture(arch, L.spectrum))