+++ SpeedyTransforms/test/runtests.jl
@@ -6,6 +6,7 @@ include("spectral_transform.jl")
 include("spectral_gradients.jl")
 include("power_spectrum.jl")
 include("resolutions.jl")
+include("matrix_transform.jl")
 
 # this will load Enzyme, EnzymeTestUtils
 if VERSION < v"1.12"